Move `rand_tx` to `test_util`
