+++ zkevm-circuits/src/pi_circuit.rs
@@ -769,55 +769,14 @@ impl<F: Field, const MAX_TXS: usize, const MAX_CALLDATA: usize> Circuit<F>
 mod pi_circuit_test {
     use super::*;
 
-    use ethers_core::{
-        types::{NameOrAddress, TransactionRequest},
-        utils::keccak256,
-    };
-    use ethers_signers::{LocalWallet, Signer};
+    use crate::test_util::rand_tx;
     use halo2_proofs::{
         dev::{MockProver, VerifyFailure},
         pairing::bn256::Fr,
     };
     use pretty_assertions::assert_eq;
-    use rand::{CryptoRng, Rng, SeedableRng};
+    use rand::SeedableRng;
     use rand_chacha::ChaCha20Rng;
-
-    // Copied from tx_circuit
-    fn rand_tx<R: Rng + CryptoRng>(mut rng: R, chain_id: u64) -> Transaction {
-        let wallet0 = LocalWallet::new(&mut rng).with_chain_id(chain_id);
-        let wallet1 = LocalWallet::new(&mut rng).with_chain_id(chain_id);
-        let from = wallet0.address();
-        let to = wallet1.address();
-        let data = b"hello";
-        let tx = TransactionRequest::new()
-            .from(from)
-            .to(to)
-            .nonce(3)
-            .value(1000)
-            .data(data)
-            .gas(500_000)
-            .gas_price(1234);
-        let tx_rlp = tx.rlp(chain_id);
-        let sighash = keccak256(tx_rlp.as_ref()).into();
-        let sig = wallet0.sign_hash(sighash, true);
-        let to = tx.to.map(|to| match to {
-            NameOrAddress::Address(a) => a,
-            _ => unreachable!(),
-        });
-        Transaction {
-            from: tx.from.unwrap(),
-            to,
-            gas_limit: tx.gas.unwrap(),
-            gas_price: tx.gas_price.unwrap(),
-            value: tx.value.unwrap(),
-            call_data: tx.data.unwrap(),
-            nonce: tx.nonce.unwrap(),
-            v: sig.v,
-            r: sig.r,
-            s: sig.s,
-            ..Transaction::default()
-        }
-    }
 
     /// Compute the raw_public_inputs column from the verifier's perspective.
     fn raw_public_inputs_col<F: Field, const MAX_TXS: usize, const MAX_CALLDATA: usize>(
