+++ zkevm-circuits/src/test_util.rs
@@ -1,9 +1,14 @@
 use crate::{evm_circuit::witness::Block, state_circuit::StateCircuit};
 use bus_mapping::mock::BlockData;
-use eth_types::geth_types::GethData;
+use eth_types::geth_types::{GethData, Transaction};
+use ethers_core::types::{NameOrAddress, TransactionRequest};
+use ethers_core::utils::keccak256;
+use ethers_signers::{LocalWallet, Signer};
 use halo2_proofs::dev::{MockProver, VerifyFailure};
 use halo2_proofs::pairing::bn256::Fr;
 use mock::TestContext;
+use rand::{CryptoRng, Rng};
+use strum::IntoEnumIterator;
 
 #[cfg(test)]
 #[ctor::ctor]
@@ -70,4 +75,40 @@ pub fn test_circuits_using_witness_block(
     }
 
     Ok(())
+}
+
+pub fn rand_tx<R: Rng + CryptoRng>(mut rng: R, chain_id: u64) -> Transaction {
+    let wallet0 = LocalWallet::new(&mut rng).with_chain_id(chain_id);
+    let wallet1 = LocalWallet::new(&mut rng).with_chain_id(chain_id);
+    let from = wallet0.address();
+    let to = wallet1.address();
+    let data = b"hello";
+    let tx = TransactionRequest::new()
+        .from(from)
+        .to(to)
+        .nonce(3)
+        .value(1000)
+        .data(data)
+        .gas(500_000)
+        .gas_price(1234);
+    let tx_rlp = tx.rlp(chain_id);
+    let sighash = keccak256(tx_rlp.as_ref()).into();
+    let sig = wallet0.sign_hash(sighash, true);
+    let to = tx.to.map(|to| match to {
+        NameOrAddress::Address(a) => a,
+        _ => unreachable!(),
+    });
+    Transaction {
+        from: tx.from.unwrap(),
+        to,
+        gas_limit: tx.gas.unwrap(),
+        gas_price: tx.gas_price.unwrap(),
+        value: tx.value.unwrap(),
+        call_data: tx.data.unwrap(),
+        nonce: tx.nonce.unwrap(),
+        v: sig.v,
+        r: sig.r,
+        s: sig.s,
+        ..Transaction::default()
+    }
 }
